requirements.txt
pytorch >= 2.2.0
tqdm
simple-parsing
munch
pyyaml